refactor: changed package names and package folder for compatible option and category config
